fix: support clang build with aarch64-unknown-windows-msvc target
libtl could be built using clang with -target aarch64-unknown-windows-msvc
for edk2 [0] but it makes a build failure:

1. INFO - /__w/1/s/ArmPkg/Library/ArmTransferListLib/libtl/include/transfer_list.h:126:56: note: expression evaluates to '9 == 8'
   INFO -   126 | LIBTL_STATIC_ASSERT(sizeof(struct transfer_list_entry) == 0x8U,

2. INFO - /home/yeoyun01/work/uefi/edk2/ArmPkg/Library/ArmTransferListLib/libtl/src/generic/transfer_list.c:51:8: error: cast to smaller integer type 'unsigned long' from 'void *' [-Werror,-Wvoid-pointer-to-int-cast]
   INFO -    51 |                      (unsigned long)transfer_list_entry_data(te));

To address this,
  (1) define hdr_size in transfer_list_entry with bitfields so that
      make hdr_size be deplyed on the same uint32_t with tag_id
      regardless of compiler.

  (2) cast uintptr_t first.

Link: https://github.com/tianocore/edk2/pull/12346 [0]

+++ src/generic/transfer_list.c
@@ -48,7 +48,7 @@ void transfer_entry_dump(struct transfer_list_entry *te)
 		info("hdr_size   0x%x\n", te->hdr_size);
 		info("data_size  0x%x\n", te->data_size);
 		info("data_addr  0x%lx\n",
-		     (unsigned long)transfer_list_entry_data(te));
+		     (unsigned long)(uintptr_t)transfer_list_entry_data(te));
 	}
 }
